requirements.txt
astral==3.2
astropy==7.0.1
astropy-iers-data==0.2025.3.10.0.29.26
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.57.0
idna==3.10
jplephem==2.22
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
pyerfa==2.0.1.5
pyparsing==3.2.3
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sgp4==2.24
six==1.17.0
skyfield==1.53
tzdata==2025.2
urllib3==2.3.0
VedAstro==1.23.18
